epv_valuation_model/requirements.txt
pandas>=1.3.0
numpy>=1.20.0
yfinance>=0.2.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
# plotly>=5.0.0 # Optional: for more interactive plots, uncomment if you plan to use it
python-dotenv>=0.19.0
# Optional: For more advanced financial modeling if needed later
# openpyxl # For reading/writing Excel files if you decide to use them
# statsmodels # For more advanced statistical modeling

